Prepare Data for Within-Subject Mediation Analysis

Prepares a dataset for within-subject mediation analysis by calculating difference scores and centered average scores for specified mediators and the outcome variable. The function ensures that the input data meets the necessary requirements and generates new variables required for subsequent mediation analysis.

Usage

PrepareData(data, M_before, M_after, Y_before, Y_after)

Arguments

data

A data frame containing the raw dataset with mediator and outcome variables.

M_before

A character vector of column names representing mediators "before" the intervention.

M_after

A character vector of column names representing mediators "after" the intervention. Must match the length of M_before.

Y_before

A character string representing the column name of the outcome variable "before" the intervention.

Y_after

A character string representing the column name of the outcome variable "after" the intervention.

Value

A data frame containing the following columns:

Ydiff: Difference score for the outcome variable.

M1diff, M2diff, ...: Difference scores for each mediator.

M1avg, M2avg, ...: Centered average scores for each mediator.

Details

This function processes raw data to create variables essential for within-subject mediation analysis. It performs the following operations:

Difference scores: Calculates the difference between "before" and "after" values for the outcome variable (Ydiff) and each mediator (Mdiff).

Centered average scores: Computes the centered average of "before" and "after" values for each mediator (Mavg), providing a measure of their mean level relative to their centered baseline.

Input validation: Checks that the number of "before" and "after" mediators match, and ensures all specified variables exist in the dataset.

This function is a prerequisite for generating structural equation modeling (SEM) syntax and conducting mediation analysis.

See also

Examples

# Example raw data
data <- data.frame(
  M1_before = rnorm(100), M1_after = rnorm(100),
  M2_before = rnorm(100), M2_after = rnorm(100),
  Y_before = rnorm(100), Y_after = rnorm(100)
)

# Prepare the dataset
prepared_data <- PrepareData(
  data = data,
  M_before = c("M1_before", "M2_before"),
  M_after = c("M1_after", "M2_after"),
  Y_before = "Y_before",
  Y_after = "Y_after"
)

head(prepared_data)
#>         Ydiff     M1diff     M2diff       M1avg      M2avg
#> 1 -0.75302767  0.7476962 -1.1761482 -0.17212362  0.5366268
#> 2 -2.20611377 -0.8840713  0.2103783  0.63375995 -0.1059446
#> 3 -3.17466365  1.8665398 -0.2529633 -0.26547750  0.9412877
#> 4  0.89805361 -0.7082856  2.1920464  1.07507793 -1.1536476
#> 5 -0.88332608 -0.9247141  0.4773740  0.40203632 -0.7559416
#> 6 -0.00701146 -0.7242002  1.7423818 -0.07876692  1.0755121