Print Method for wsMed Objects

Provides a comprehensive summary of results from a wsMed object, including:

Input and computed variables with sample size.

Model fit indices, regression paths, and variance estimates.

Total, direct, and indirect effects with pairwise contrasts.

Moderation effects and Monte Carlo confidence intervals for raw and standardized estimates (if applicable).

Diagnostic notes for bootstrapping, imputation, and analysis parameters.

The output is formatted for clarity, ensuring an intuitive presentation of mediation analysis results, including dynamic confidence intervals, moderation keys, and pre-post coefficients.

Usage

# S3 method for class 'wsMed'
print(x, level = 0.95, digits = 3, ...)

Arguments

x

A wsMed object containing the results of within-subject mediation analysis.

level

Numeric. Confidence level for the intervals (default = 0.95).

digits

Numeric. Number of digits to display in the results.

...

Additional arguments (not used currently).

Value

Invisibly returns the input wsMed object for further use.

Details

This function is specifically designed to display results from the within-subject mediation analysis conducted using the wsMed function. Key features include:

Variables:

Shows input variables (M_before, M_after, Y_before, Y_after) and computed variables like Ydiff, Mdiff, and Mavg.

Reports the sample size used in the analysis.

Model Fit Indices:

Displays SEM fit indices (e.g., Chi-square, CFI, TLI, RMSEA, SRMR) to assess model quality.

Regression Paths and Variance Estimates:

Summarizes path coefficients, intercepts, variances, and confidence intervals.

Effects:

Reports total, direct, and indirect effects with their significance.

Highlights pairwise contrasts between indirect effects for mediation paths.

Moderation Effects:

Provides moderation results for identified variables with corresponding coefficients and paths.

Monte Carlo Confidence Intervals:

Includes results for raw and standardized estimates obtained using methods such as MI or FIML.

Diagnostics:

Summarizes analysis parameters like bootstrapping, imputation settings, Monte Carlo iterations, and random seeds.

See also

Examples

# Example dataset with missing values
data(example_data)
set.seed(123)
example_dataN <- mice::ampute(
  data = example_data,
  prop = 0.1
)$amp

# Perform within-subject mediation analysis
result1 <- wsMed(
  data = example_dataN,
  M_before = c("A1", "B1"),
  M_after = c("A2", "B2"),
  Y_before = "C1",
  Y_after = "C2",
  form = "P",
  Na = "MI",
  standardized = TRUE,
  bootstrap = 1000,
  iseed = 123,
  se = "boot",
  R = 20000L,  # Monte Carlo repetitions
  alpha = c(0.001, 0.01, 0.05),  # Significance levels
  m = 5,  # Number of imputations
  method = "pmm",  # Imputation method
  decomposition = "eigen",
  pd = TRUE,
  tol = 1e-06,
  seed = 123,
  alphastd = c(0.01, 0.05)
)
#> 
#>  iter imp variable
#>   1   1  A3  B1  B2  C1  C3  D1
#>   1   2  A3  B1  B2  C1  C3  D1
#>   1   3  A3  B1  B2  C1  C3  D1
#>   1   4  A3  B1  B2  C1  C3  D1
#>   1   5  A3  B1  B2  C1  C3  D1
#>   2   1  A3  B1  B2  C1  C3  D1
#>   2   2  A3  B1  B2  C1  C3  D1
#>   2   3  A3  B1  B2  C1  C3  D1
#>   2   4  A3  B1  B2  C1  C3  D1
#>   2   5  A3  B1  B2  C1  C3  D1
#>   3   1  A3  B1  B2  C1  C3  D1
#>   3   2  A3  B1  B2  C1  C3  D1
#>   3   3  A3  B1  B2  C1  C3  D1
#>   3   4  A3  B1  B2  C1  C3  D1
#>   3   5  A3  B1  B2  C1  C3  D1
#>   4   1  A3  B1  B2  C1  C3  D1
#>   4   2  A3  B1  B2  C1  C3  D1
#>   4   3  A3  B1  B2  C1  C3  D1
#>   4   4  A3  B1  B2  C1  C3  D1
#>   4   5  A3  B1  B2  C1  C3  D1
#>   5   1  A3  B1  B2  C1  C3  D1
#>   5   2  A3  B1  B2  C1  C3  D1
#>   5   3  A3  B1  B2  C1  C3  D1
#>   5   4  A3  B1  B2  C1  C3  D1
#>   5   5  A3  B1  B2  C1  C3  D1
#> Class: mids
#> Number of multiple imputations:  5 
#> Imputation methods:
#>    A1    A2    A3    B1    B2    B3    C1    C2    C3    D1    D2    D3 
#>    ""    "" "pmm" "pmm" "pmm"    "" "pmm"    "" "pmm" "pmm"    ""    "" 
#> PredictorMatrix:
#>    A1 A2 A3 B1 B2 B3 C1 C2 C3 D1 D2 D3
#> A1  0  0  0  0  0  0  0  0  0  0  0  0
#> A2  0  0  0  0  0  0  0  0  0  0  0  0
#> A3  1  1  0  1  1  1  1  1  1  1  1  1
#> B1  1  1  1  0  1  1  1  1  0  1  1  1
#> B2  1  1  1  1  0  1  1  1  0  1  1  1
#> B3  0  0  0  0  0  0  0  0  0  0  0  0

# Print the results
print(result1)
#> 
#> *************** VARIABLES ***************
#> Variables:
#> Y =  C2 C1 
#> M1 =  A2 A1 
#> M2 =  B2 B1 
#> 
#> Computed Variables:
#> 
#> 
#> | Variable|                  Formula|
#> |--------:|------------------------:|
#> |    Ydiff|                  C2 - C1|
#> |   M1diff|                  A2 - A1|
#> |   M2diff|                  B2 - B1|
#> |    M1avg| ( A2 + A1 ) / 2 Centered|
#> |    M2avg| ( B2 + B1 ) / 2 Centered|
#> 
#> Sample Size:  100 
#> Confidence Level:  95 %
#> 
#> 
#> *************** MODEL FIT INDICES ***************
#> 
#> 
#> |            Measure|  Value|
#> |------------------:|------:|
#> |         Chi-Square| 10.742|
#> | Degrees of Freedom|  5.000|
#> |            p-Value|  0.057|
#> |                CFI|  0.000|
#> |                TLI| -1.384|
#> |              RMSEA|  0.109|
#> |     RMSEA Lower CI|  0.000|
#> |     RMSEA Upper CI|  0.199|
#> |               SRMR|  0.074|
#> 
#> *************** REGRESSION PATHS, INTERCEPTS AND VARIANCES ***************
#> 
#> 
#> |           Path| Label| Estimate|    SE|      z| P.value|   LLCI|  ULCI|
#> |--------------:|-----:|--------:|-----:|------:|-------:|------:|-----:|
#> | Ydiff ~ M1diff|    b1|   -0.038| 0.093| -0.404|   0.686| -0.221| 0.145|
#> | Ydiff ~ M2diff|    b2|   -0.115| 0.090| -1.272|   0.204| -0.292| 0.062|
#> |  Ydiff ~ M1avg|    d1|   -0.051| 0.095| -0.541|   0.589| -0.237| 0.135|
#> |  Ydiff ~ M2avg|    d2|   -0.063| 0.084| -0.748|   0.455| -0.229| 0.102|
#> 
#> 
#> | Intercept| Label| Estimate|    SE|      z| P.value|   LLCI|  ULCI|
#> |---------:|-----:|--------:|-----:|------:|-------:|------:|-----:|
#> |   Ydiff~1|    cp|    0.017| 0.017|  1.004|   0.315| -0.016| 0.050|
#> |  M1diff~1|    a1|   -0.030| 0.018| -1.706|   0.088| -0.065| 0.005|
#> |  M2diff~1|    a2|    0.016| 0.018|  0.843|   0.399| -0.021| 0.052|
#> |   M1avg~1|      |   -0.008| 0.018| -0.448|   0.654| -0.044| 0.028|
#> |   M2avg~1|      |   -0.000| 0.020| -0.005|   0.996| -0.040| 0.040|
#> 
#> 
#> |       Variance| Estimate|    SE|     z| P.value|  LLCI|  ULCI|
#> |--------------:|--------:|-----:|-----:|-------:|-----:|-----:|
#> |   Ydiff~~Ydiff|    0.026| 0.004| 6.964|   0.000| 0.019| 0.033|
#> | M1diff~~M1diff|    0.031| 0.004| 6.964|   0.000| 0.022| 0.040|
#> | M2diff~~M2diff|    0.033| 0.005| 6.964|   0.000| 0.024| 0.042|
#> |   M1avg~~M1avg|    0.032| 0.005| 6.964|   0.000| 0.023| 0.041|
#> |   M2avg~~M2avg|    0.041| 0.006| 6.964|   0.000| 0.029| 0.052|
#> 
#> 
#> *************** TOTAL AND DIRECT EFFECT ***************
#> 
#> 
#> |          Name| Effect|    SE|     z|     p|   LLCI|  ULCI|
#> |-------------:|------:|-----:|-----:|-----:|------:|-----:|
#> |  Total effect|  0.016| 0.017| 0.975| 0.329| -0.016| 0.049|
#> | Direct effect|  0.017| 0.017| 1.004| 0.315| -0.016| 0.050|
#> 
#> 
#> *************** INDIRECT EFFECTS ***************
#> 
#> 
#> |      Name| Effect|    SE|   LLCI|  ULCI|
#> |---------:|------:|-----:|------:|-----:|
#> |      ind1|  0.001| 0.003| -0.005| 0.007|
#> |      ind2| -0.002| 0.003| -0.007| 0.003|
#> | total ind| -0.001| 0.004| -0.008| 0.007|
#> 
#> 
#> | Ind  |         Path         |
#> |:----:|:--------------------:|
#> | ind1 | X -> M1diff -> Ydiff |
#> | ind2 | X -> M2diff -> Ydiff |
#> 
#> 
#> *************** CONTRAST INDIRECT EFFECTS ***************
#> 
#> 
#> |        Name| Effect|    SE|   LLCI|  ULCI|
#> |-----------:|------:|-----:|------:|-----:|
#> | ind1 - ind2|  0.003| 0.004| -0.005| 0.011|
#> 
#> 
#> *************** MODERATION EFFECTS of X ***************
#> 
#> 
#> | Name| Effect|    SE|      z|     p|   LLCI|  ULCI|
#> |----:|------:|-----:|------:|-----:|------:|-----:|
#> |   d1| -0.051| 0.095| -0.541| 0.589| -0.237| 0.135|
#> |   d2| -0.063| 0.084| -0.748| 0.455| -0.229| 0.102|
#> 
#> 
#> | Coefficient |      Path       |
#> |:-----------:|:---------------:|
#> |     d1      | M1diff -> Ydiff |
#> |     d2      | M2diff -> Ydiff |
#> 
#> 
#> *************** PRE-POST COEFFICIENTS ***************
#> 
#> 
#> |  Name| Effect|    SE|      z|     p|   LLCI|  ULCI|
#> |-----:|------:|-----:|------:|-----:|------:|-----:|
#> | X1_b1| -0.063| 0.105| -0.605| 0.545| -0.269| 0.142|
#> | X0_b1| -0.012| 0.105| -0.115| 0.908| -0.217| 0.193|
#> | X1_b2| -0.146| 0.100| -1.468| 0.142| -0.342| 0.049|
#> | X0_b2| -0.083| 0.100| -0.835| 0.404| -0.279| 0.112|
#> 
#> 
#> | Coefficient |       Path       |
#> |:-----------:|:----------------:|
#> |     b1      | M1diff -> Ydiff  |
#> |     b2      | M2diff -> Ydiff  |
#> |     b12     | M1diff -> M2diff |
#> 
#> 
#> *************** MONTE CARLO CONFIDENCE INTERVALS (MI) ***************
#> 
#> 
#> |      Parameter| Estimate|    SE|   0.1%|   0.5%|   2.5%|  97.5%|  99.5%|  99.9%|
#> |--------------:|--------:|-----:|------:|------:|------:|------:|------:|------:|
#> |  direct effect|    0.016| 0.017| -0.039| -0.028| -0.017|  0.049|  0.059|  0.070|
#> |             b1|   -0.035| 0.092| -0.338| -0.269| -0.214|  0.145|  0.198|  0.250|
#> |             b2|   -0.117| 0.090| -0.416| -0.348| -0.297|  0.057|  0.115|  0.173|
#> |             d1|   -0.063| 0.092| -0.358| -0.301| -0.241|  0.118|  0.175|  0.236|
#> |             d2|   -0.068| 0.084| -0.352| -0.288| -0.233|  0.096|  0.146|  0.198|
#> |             a1|   -0.027| 0.018| -0.087| -0.073| -0.061|  0.007|  0.018|  0.033|
#> |             a2|    0.017| 0.018| -0.042| -0.029| -0.018|  0.052|  0.064|  0.076|
#> |   Ydiff~~Ydiff|    0.026| 0.004|  0.014|  0.016|  0.019|  0.033|  0.035|  0.039|
#> | M1diff~~M1diff|    0.031| 0.004|  0.016|  0.020|  0.022|  0.039|  0.042|  0.045|
#> | M2diff~~M2diff|    0.032| 0.005|  0.018|  0.021|  0.023|  0.041|  0.045|  0.048|
#> |   M1avg~~M1avg|    0.034| 0.000|  0.034|  0.034|  0.034|  0.034|  0.034|  0.034|
#> |   M1avg~~M2avg|    0.011| 0.000|  0.011|  0.011|  0.011|  0.011|  0.011|  0.011|
#> |   M2avg~~M2avg|    0.040| 0.000|  0.040|  0.040|  0.040|  0.040|  0.040|  0.040|
#> |        M1avg~1|   -0.000| 0.000| -0.000| -0.000| -0.000| -0.000| -0.000| -0.000|
#> |        M2avg~1|   -0.000| 0.000| -0.000| -0.000| -0.000| -0.000| -0.000| -0.000|
#> |           ind1|    0.001| 0.003| -0.012| -0.008| -0.005|  0.008|  0.012|  0.017|
#> |           ind2|   -0.002| 0.003| -0.019| -0.013| -0.010|  0.003|  0.005|  0.008|
#> |      total ind|   -0.001| 0.004| -0.021| -0.015| -0.010|  0.007|  0.011|  0.016|
#> |   total effect|    0.015| 0.017| -0.040| -0.028| -0.018|  0.048|  0.058|  0.069|
#> |      ind1-ind2|    0.003| 0.004| -0.012| -0.008| -0.005|  0.013|  0.017|  0.024|
#> |          X1_b1|   -0.067| 0.103| -0.410| -0.329| -0.270|  0.134|  0.195|  0.255|
#> |          X0_b1|   -0.003| 0.102| -0.341| -0.264| -0.200|  0.195|  0.254|  0.312|
#> |          X1_b2|   -0.151| 0.099| -0.475| -0.411| -0.346|  0.043|  0.103|  0.165|
#> |          X0_b2|   -0.083| 0.100| -0.420| -0.343| -0.278|  0.111|  0.170|  0.237|
#> 
#> 
#> *************** MONTE CARLO CONFIDENCE INTERVALS (STANDARDIZED) ***************
#> 
#> 
#> |      Parameter| Estimate|    SE|   0.5%|   2.5%|  97.5%|  99.5%|
#> |--------------:|--------:|-----:|------:|------:|------:|------:|
#> |             b1|   -0.030| 0.097| -0.287| -0.225|  0.153|  0.210|
#> |             b2|   -0.137| 0.097| -0.368| -0.315|  0.063|  0.126|
#> |             d1|   -0.057| 0.101| -0.326| -0.264|  0.131|  0.195|
#> |             d2|   -0.094| 0.101| -0.339| -0.276|  0.116|  0.179|
#> |   Ydiff~~Ydiff|    0.965| 0.042|  0.787|  0.830|  0.989|  0.995|
#> | M1diff~~M1diff|    1.000| 0.000|  1.000|  1.000|  1.000|  1.000|
#> | M2diff~~M2diff|    1.000| 0.000|  1.000|  1.000|  1.000|  1.000|
#> |   M1avg~~M1avg|    1.000| 0.000|  1.000|  1.000|  1.000|  1.000|
#> |   M1avg~~M2avg|    0.284| 0.000|  0.284|  0.284|  0.284|  0.284|
#> |   M2avg~~M2avg|    1.000| 0.000|  1.000|  1.000|  1.000|  1.000|
#> |           ind1|    0.005| 0.000| -0.000| -0.000| -0.000| -0.000|
#> |           ind2|   -0.014| 0.000| -0.000| -0.000| -0.000| -0.000|
#> |      total ind|   -0.010| 0.000| -0.000| -0.000| -0.000| -0.000|
#> |   total effect|    0.103| 0.000|  0.000|  0.000|  0.000|  0.000|
#> |      ind1-ind2|    0.019| 0.000|  0.000|  0.000|  0.000|  0.000|
#> |          X1_b1|   -0.058| 0.110| -0.348| -0.288|  0.143|  0.206|
#> |          X0_b1|   -0.002| 0.109| -0.284| -0.211|  0.210|  0.272|
#> |          X1_b2|   -0.184| 0.108| -0.441| -0.376|  0.048|  0.119|
#> |          X0_b2|   -0.090| 0.110| -0.366| -0.302|  0.129|  0.196|
#> 
#> 
#> *************** IMPUTATION AND MONTE CARLO NOTES ***************
#> 
#> Number of imputations (m):  5 
#> Imputation method:  pmm 
#> Random seed:  123 
#> Number of Monte Carlo repetitions (R):  20000 
#> Decomposition method for covariance matrices:  eigen 
#> Check positive definiteness of covariance matrices:  Yes 
#> Tolerance for positive definiteness checks :  1e-06 
#> Significance levels for confidence intervals:  0.001, 0.01, 0.05 
#> Significance levels for standardized confidence intervals:  0.01, 0.05